[tb/LEMS-3140/group-user-input-validation] Move widget ID component parser to own file
